Eliminación de las imagenes de los inmuebles

+++ php/eliminarInmueble.php
@@ -1,9 +1,21 @@
 <?php
   $IdInmueble = $_GET['IdInmueble'];
   include('dbconnection.php');
-  $sql = "DELETE FROM Inmueble WHERE IdInmueble='".$IdInmueble."';";
+  $sql = "DELETE FROM InmuebleFoto WHERE IdInmueble='".$IdInmueble."';";
   if(mysqli_query($conn, $sql)){
-    echo "Información eliminada correctamente";
+    $sql = "DELETE FROM Visitas WHERE Inmueble='".$IdInmueble."';";
+    if(mysqli_query($conn, $sql)){
+      $sql = "DELETE FROM Inmueble WHERE IdInmueble='".$IdInmueble."';";
+      if(mysqli_query($conn, $sql)){
+        echo "Información eliminada correctamente";
+      }
+      else{
+        echo "Error al eliminar la información";
+      }
+    }
+    else{
+      echo "Error al eliminar la información";
+    }
   }
   else{
     echo "Error al eliminar la información";
